build: update e2fsprogs.spec for distro builds
Add the distro version to the RPM release number, so that it the
RPM names do not conflict.

Allow the RPM built from upstream to replace the split packages
provided by the distros.  At some point in the future it may be
desirable to also split the RPM built by this spec file, but this
bs complicated by the fact that SLES and RHEL have different splits.

  LU-19047 build: RPM should provide ldiskfsprogs-devel

  Add ldiskfsprogs-devel so that this can be used to disinguish the
  Lustre-patched e2fsprogs development library when building DKMS RPMs.

  Signed-off-by: Andreas Dilger <adilger@whamcloud.com>
  Change-Id: I27fb796dba378a0596952de50618e64a98dc34d1
  Reviewed-on: https://review.whamcloud.com/c/tools/e2fsprogs/+/59413
  Tested-by: jenkins <devops@whamcloud.com>
  Tested-by: Maloo <maloo@whamcloud.com>
  Reviewed-by: Shaun Tancheff <shaun.tancheff@hpe.com>
  Reviewed-by: Li Dongyang <dongyangli@ddn.com>

Signed-off-by: Andreas Dilger <adilger@whamcloud.com>
Change-Id: Iab8f75a7f7bd9790899a2024d91079d55501ef65