TT-177 build: add .spec file for SLES11+ packaging
Include the upstream SLES11 .spec file to ensure the packages we
build match the upstream packages.  Any later patches that change
the packaging should patch the .spec file appropriately.

Add in the SLES-specific patches, excluding the replacement de.po
file, since the original SLES11 de.po file is only against 1.41.4,
and is missing a large number of changes to the translated messages
related to 64-bit format specifiers.

Always skip m_hugefile and t_replay_and_set tests, building rpms in
an ext3 environment will make the tests fail.

Need to add a line in the SUSE spec file for Provides: ldiskfsprogs.
This is present in the RHEL spec file and is needed to resolve
dependencies in lustre server rpms at rpm install time.

Ensure that the libcom_err and libext2fs2 packages installed match the
other packages built from e2fsprogs so that modifications to those
libraries (e.g. project quota) are available.

Changes in output from the lsb_release command in SLES 12
resulted in using the wrong .spec file in builds.
This patch corrects the problem for old and new versions.

  LU-19047 build: RPM should provide ldiskfsprogs-devel

  Add ldiskfsprogs-devel so that this can be used to disinguish the
  Lustre-patched e2fsprogs development library when building DKMS RPMs.

  Signed-off-by: Andreas Dilger <adilger@whamcloud.com>
  Change-Id: I27fb796dba378a0596952de50618e64a98dc34d1
  Reviewed-on: https://review.whamcloud.com/c/tools/e2fsprogs/+/59413
  Tested-by: jenkins <devops@whamcloud.com>
  Tested-by: Maloo <maloo@whamcloud.com>
  Reviewed-by: Shaun Tancheff <shaun.tancheff@hpe.com>
  Reviewed-by: Li Dongyang <dongyangli@ddn.com>

Signed-off-by: Bob Glossman <bob.glossman@intel.com>
Signed-off-by: Andreas Dilger <andreas.dilger@intel.com>
Change-Id: I783d58bd78d7c4c66cc85ec5557ae1aaf64016ba