e2fsck: add Lustre lfsck tool
The lfsck tool, in conjunction with e2fsck, build a DB4 database
of all the inodes and objects on the MDT and OST filesystems.

The lfsck tool combines the databases on the Lustre client,
and can verify that all of the objects referenced by inodes
exist, are not referenced by two inodes, and have a parent
inode.

  LU-266 e2fsck: regenerate LAST_ID file

  e2fsck should be able to regenerate the LAST_ID file if it gets
  corrupted.  This patch will create a new LAST_ID file if it was
  deleted, and removes the unnecessary lfsck_get_last_id function.
  The last_id is then set as before in e2fsck_pass6_ost to be either
  the max objid on disk, or the MDS' max ost id, whichever is larger.

  Reported-by: Bernd Schubert <aakef@fastmail.fm>
  Signed-off-by: Kit Westneat <kwestneat@ddn.com>
  Change-Id: Ic5396da000909b826b76da2fd5a0b5ce88b06944

Signed-off-by: Andreas Dilger <adilger@whamcloud.com>