LU-266 e2fsck: regenerate LAST_ID file
e2fsck should be able to regenerate the LAST_ID file if it gets corrupted.
This patch will create a new LAST_ID file if it was deleted, and removes the
unnecessary lfsck_get_last_id function. The last_id is then set as before in
e2fsck_pass6_ost to be either the max objid on disk, or the MDS' max ost id,
whichever is larger.

Reported-by: Bernd Schubert <aakef@fastmail.fm>
Signed-off-by: Kit Westneat <kwestneat@ddn.com>
Change-Id: Ic5396da000909b826b76da2fd5a0b5ce88b06944