build: run long tests for RPMs, and run them in parallel
Change the "make check" in the RPM .spec.in files to run all of the
tests ("fullcheck" vs. only "check", which now excludes slow tests
by default), but change the .spec file to run the tests in parallel.

This allows long-running tests to start and wait in the background,
while allowing some short-running to actually do work in the meantime.

Signed-off-by: Andreas Dilger <adilger@whamcloud.com>
Change-Id: I36d8d64907c58ceb25bb955edeeac4e1e53ebbe5
Reviewed-on: https://review.whamcloud.com/33570
Tested-by: Jenkins
Tested-by: Maloo <maloo@whamcloud.com>
Reviewed-by: Li Dongyang <dongyangli@ddn.com>
Reviewed-by: Yang Sheng <ys@whamcloud.com>