LU-13197 misc: handle large files with filefrag
Avoid overflowing the column-width calc when printing files over 4B blocks.

Document the availability of [KMG] suffixes for the "-b <blocksize>" option.
The blocksize is limited to at most 1GiB blocksize to avoid shifting all of
the extents down to zero GB in size.  Even the use of 1GB blocksize is a bit
tenuous, but some filesystems may multi-GB extents beyond ext4.

Change-Id: I003a721a3d199b1cb8b3912236328eb80b34f0fe
Signed-off-by: Andreas Dilger <adilger@dilger.ca>
Reviewed-on: https://review.whamcloud.com/37439
Tested-by: jenkins <devops@whamcloud.com>
Tested-by: Maloo <maloo@whamcloud.com>
Reviewed-by: Wang Shilong <wshilong@ddn.com>
Reviewed-by: Andreas Dilger <adilger@whamcloud.com>