LU-13197 e2fsck: fix overflow if more than 4B inodes
Even though we don't have support for filesystems with over 4B inodes
in the current e2fsprogs, this may happen in the future.  There are
latent overflow bugs when calculating the number of inodes in the
filesystem that may as well be fixed now, rather than waiting for it
to be hit in the field at some point.

Signed-off-by: Andreas Dilger <adilger@dilger.ca>
Change-Id: I336ca04fe8f944dbfbe5425ce84e7d92743ebbe5
Reviewed-on: https://review.whamcloud.com/37440
Tested-by: jenkins <devops@whamcloud.com>
Tested-by: Maloo <maloo@whamcloud.com>
Reviewed-by: Wang Shilong <wshilong@ddn.com>
Reviewed-by: Andreas Dilger <adilger@whamcloud.com>