LU-13797 e2fsck: quiet 'extent tree could be narrower'
The "extent tree (at level 1) could be narrower" message is
overly verbose and raises concerns by end users, even though
it is harmless.  On the flip side, this may save only a few
hundred blocks in the filesystem for a short period of time,
so there is relatively little benefit to be had.

Disable the extent optimization step in e2fsck by default by
adding the "no_optimize_extents" option to e2fsck.conf.

Change-Id: I0cee9407ef0653c3b946908513f3429816b89548
Signed-off-by: Andreas Dilger <adilger@whamcloud.com>
Reviewed-on: https://review.whamcloud.com/40952
Tested-by: jenkins <devops@whamcloud.com>
Tested-by: Maloo <maloo@whamcloud.com>
Reviewed-by: Wang Shilong <wshilong@whamcloud.com>