LU-15330 e2fsck: no parent lookup in disconnected dir
Don't call ext2fs_get_pathname() to do a name lookup for a
disconnected directory, since the directory block traversal
in pass1 has already scanned all of the leaf blocks and not
found the relevant entry.  If it had been found the directory
would not be disconnected in pass3.

Instead, print the ".." directory name in the prompt directly.
This avoids a potentially lengthy directory traversal for each
disconnected entry if the previous parent is large that never
finds the entry.

Signed-off-by: Andreas Dilger <adilger@whamcloud.com>
Change-Id: If17a92689f24f365ca1fbe5c837e7d5f383ebbe5
Reviewed-on: https://review.whamcloud.com/45785
Tested-by: Maloo <maloo@whamcloud.com>
Tested-by: jenkins <devops@whamcloud.com>
Reviewed-by: Artem Blagodarenko <artem.blagodarenko@hpe.com>