LU-16972 e2fsck: use rb-tree to track EA reference counts
Using the sorted array to track the EA blocks and
its refs is not scalable.
When the file system has a huge number of EA blocks,
pass1 scanning could not be finished within a reasonable
time, as 95%+ of CPU time is spent in memmove() when
trying to enlarge the the sorted array.

On a file system with 20 million EA blocks on an NVMe device
pass1 time taken:
without patch:
time: 2014.78/1838.70/19.91
with patch:
time: 45.17/20.17/20.19

Change-Id: I6dc1ee3037dbf7a48deb610514af1f0e35a5a397
Signed-off-by: Li Dongyang <dongyangli@ddn.com>
Reviewed-on: https://review.whamcloud.com/c/tools/e2fsprogs/+/51729
Tested-by: jenkins <devops@whamcloud.com>
Tested-by: Maloo <maloo@whamcloud.com>
Reviewed-by: Andreas Dilger <adilger@whamcloud.com>