LU-17117 debugfs: allow debugfs to disable MMP directly
Instead of needing to use "tune2fs -f -E clear_mmp" to clear the
MMP status, it should be possible to open and clear the MMP block
directly with debugfs in case of issues with tune2fs.

Add the "-m" option to debugfs and the "open" command to allow
opening the device and skipping the MMP status check.  Otherwise,
if sequence EXT4_MMP_SEQ_FSCK is set then debugfs will report an
error and the device cannot be not opened for write, even with -f:

    debugfs -f /dev/sdc
    MMP: e2fsck being run while trying to open /dev/sdc

Add the "clear_mmp" command to match the tune2fs extended command.
Add command aliases "mmp_clear", "mmp_dump", and "mmp_set_value"
so that they can be grouped together and found more easily.

Signed-off-by: Andreas Dilger <adilger@whamcloud.com>
Change-Id: I580ba4f30c366b0d0e75596c80cef2892c408fb0
Reviewed-on: https://review.whamcloud.com/c/tools/e2fsprogs/+/52380
Tested-by: jenkins <devops@whamcloud.com>
Tested-by: Maloo <maloo@whamcloud.com>
Reviewed-by: Li Dongyang <dongyangli@ddn.com>