LU-19105 e2fsck: support duplicate dentry rename with dirdata
mutate_name() is used to generate a new name in case of duplicate dir
entries. The rename is done in-place without modifying the record
length. The record length is 4 bytes aligned. So mutate_name() uses
the padding bytes to expand the name length.

If dirdata is present in the record, the rename operation may
overwrite the dirdata saved after dir entry name.

If the name length needs to be expanded, this change moves the dirdata
before renaming.

Add the regression test f_dirdata_dup_de to verify duplicate entries
fix with dirdata.

Fixes: c06619d91 ("e2fsck: add support for dirdata feature")
Signed-off-by: Etienne AUJAMES <eaujames@ddn.com>
Change-Id: I2fcd93a0460bab1371d0621279a80d1994b26d2a
Reviewed-on: https://review.whamcloud.com/c/tools/e2fsprogs/+/59821
Tested-by: jenkins <devops@whamcloud.com>
Tested-by: Maloo <maloo@whamcloud.com>
Reviewed-by: Andreas Dilger <adilger@whamcloud.com>
Reviewed-by: Li Dongyang <dongyangli@ddn.com>