LU-18818 build: add -std=gnu11 cflag for el7 builds
Due to the configure script is stored on version control
system, sometimes when running make rpm, autoconf decides
to regen configure script.

This creates an issue for el7 builds as the autoconf
on that distro is too old to generate a configure script
that checks if the compiler needs -std=gnu11 flag, and
missing that compiler flag results in build failures.

Signed-off-by: Li Dongyang <dongyangli@ddn.com>
Change-Id: I343a44d7247d44a6d03ac3aa5a629576c7b6c93b
Reviewed-on: https://review.whamcloud.com/c/tools/e2fsprogs/+/62169
Reviewed-by: Andreas Dilger <adilger@dilger.ca>
Tested-by: jenkins <devops@whamcloud.com>
Tested-by: Maloo <maloo@whamcloud.com>